fixed the filtering problem
